Pass around the infill angle and bridge angle as double Made the bridge tests work with double. Got the largest convex side to be the side used for bridge.

+++ Settings.cs
@@ -106,7 +106,7 @@ namespace MatterHackers.MatterSlice
         public double infillExtendIntoPerimeter;
         public int infillExtendIntoPerimeter_um { get { return (int)(infillExtendIntoPerimeter * 1000); } }
 
-        public int infillStartingAngle;
+        public double infillStartingAngle;
 
         [SettingDescription("How far from objects the first skirt loop should be, in millimeters.")]
         public int skirtDistanceFromObject;
